+++ actor/Cargo.toml
@@ -20,7 +20,6 @@ crossbeam-queue.workspace = true
 futures.workspace = true
 loom = { workspace = true, optional = true }
 parking_lot.workspace = true
-tokio = { workspace = true, features = ["sync"] }
 
 [dev-dependencies]
 commonware-utils.workspace = true
